Add coq-bedrock
This is the bulk of bedrock (all the targets at the tip of master that
build with 8.4).  In trunk, it takes a bit over 21 hours of user time to
build on my machine.  There are still some admits, because the porting
is a work in progress, but I think there's enough now to make it worth
benchmarking (when you're willing to wait 21 hours).