Update the various versions of HoTT for a bit more consistency
I'm not sure why 8.7.dev was pointing at @ejgallego's repo.

I've copied the author name from the 8.6.dev file; I'm not sure it's
right, but UniMath is definitely wrong.

I've updated the issue tracker.

I'm not sure why the build script is different for the .dev version than
for the other versions, so I haven't touched it.