coq-performance-tests: More restrictive versions
There are some new files that only work with Coq 8.8 and later.  I guess in the future we can start making branches for older versions of Coq, but I'm disinclined to make a separate package for 8.5--8.7, given that it would only contain one file, and that this repo is only used for benchmarking anyway.