Sync coq/coqide dev packages with (safer, slower) release settings
These opam packages are currently used as templates for releases.
See discussion at ocaml/opam-repository#16908,
https://github.com/ocaml/opam-repository/pull/16876#issuecomment-665598798
and
https://github.com/coq/coq/issues/11178#issuecomment-665589620.

Only done for 8.12 and dev, which will be used in the future.

The resulting packages are safer and slower (for flambda users); fixing that
will require a more proper fix (such as ocaml/opam-repository#16887).