Updated repo link and package name
The repo is LogicalAtomist/principia, not .../principia-pl (although the package only covers Principia's propositional logic). Since extensions of this package to Principia's quantification theory, class theory, and logic of relations are planned, I suppose it may be good to stick with the Principia repo and merely update this package accordingly.

If it is preferable to keep different parts of Principia in separate packages (although I am not sure why that would be), then I can create separate repos for each one. Thanks kindly!