Reduce resource consumption for continuous integration.
This commit splits the setup into two stages. During the first stage, only
"opam lint" is run, and "--check-upstream" is passed to it. Indeed, there
is no point in starting a bunch of runners if the opam files are broken
and/or the upstream tarball is not available.

The traditional build happens during second stage. This commit removes
half of the tested OCaml compilers. The remaining ones are 4.07.1, 4.11.2,
4.14.0, and any.

Finally, the json artifact is set to expire after one week rather than one
year. Since the one corresponding to the top of master is kept
indefinitely, there is no point in keeping the other artifacts around for
a long time.